move SiStripCondObjectRepresent from interface to plugins directory
