+++ CondCore/SiStripPlugins/plugins/SiStripApvGain_PayloadInspector.cc
@@ -21,8 +21,8 @@
 
 // auxilliary functions
 #include "CondCore/SiStripPlugins/interface/SiStripPayloadInspectorHelper.h"
-#include "CondCore/SiStripPlugins/interface/SiStripCondObjectRepresent.h"
 #include "CalibTracker/StandaloneTrackerTopology/interface/StandaloneTrackerTopology.h"
+#include "SiStripCondObjectRepresent.h"
 
 #include <memory>
 #include <sstream>
