+++ CondCore/SiStripPlugins/plugins/SiStripCondObjectRepresent.h
